Safari-611.1.10/LayoutTests/fast/events/touch/gesture/pad-gesture-fling-expected.txt

Tests basic use of GestureFlingStart

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL actualWheelEventsOccurred should be 2. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
